crc-code/requirements.txt
azure-core==1.27.1
azure-data-tables==12.4.3
azure-functions==1.15.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
isodate==0.6.1
multidict==6.0.4
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
